ci: Publish package to npm registry
Migrates the package publishing process from GitHub Packages to the public npm registry.

This change enhances package accessibility by making it available on the primary npm registry. It updates the GitHub Actions workflow to use OIDC for authentication and explicitly targets npmjs.org.

Includes the addition of an `.npmignore` file to optimize package size and contents for npm, and an expanded `.gitignore` for general repository cleanliness.